move date settings to submodule
