+++ src/plot.rs
@@ -1,4 +1,4 @@
-use chrono::{DateTime, NaiveDateTime, Utc};
+use date_settings::LogStartDateSettings;
 use plot_util::PlotWithName;
 use serde::{Deserialize, Serialize};
 use skytem_logs::{
@@ -15,43 +15,10 @@ use play_state::{playback_update_plot, PlayState};
 use plot_visibility_config::PlotVisibilityConfig;
 
 mod axis_config;
+mod date_settings;
 mod play_state;
 mod plot_ui;
 mod plot_visibility_config;
-
-#[derive(PartialEq, Deserialize, Serialize)]
-pub struct LogStartDateSettings {
-    pub log_id: String,
-    pub original_start_date: DateTime<Utc>,
-    pub start_date: DateTime<Utc>,
-    pub clicked: bool,
-    pub tmp_date_buf: String,
-    pub err_msg: String,
-    pub new_date_candidate: Option<NaiveDateTime>,
-    pub date_changed: bool,
-}
-
-impl LogStartDateSettings {
-    pub fn new(log_id: String, start_date: DateTime<Utc>) -> Self {
-        Self {
-            log_id,
-            original_start_date: start_date,
-            start_date,
-            clicked: false,
-            tmp_date_buf: String::new(),
-            err_msg: String::new(),
-            new_date_candidate: None,
-            date_changed: false,
-        }
-    }
-
-    /// Get the offset in nanoseconds from the original start date and the adjusted start date
-    pub fn date_offset_ns(&self) -> i64 {
-        (self.start_date - self.original_start_date)
-            .num_nanoseconds()
-            .expect("Nanoseconds count exceeds capacity of i64")
-    }
-}
 
 #[allow(missing_debug_implementations)] // Legend is from egui_plot and doesn't implement debug
 #[derive(PartialEq, Deserialize, Serialize)]
@@ -220,7 +187,7 @@ impl LogPlot {
             }
 
             for settings in log_start_date_settings {
-                update_plot_dates(
+                date_settings::update_plot_dates(
                     percentage_plots,
                     to_hundreds_plots,
                     to_thousands_plots,
@@ -323,48 +290,5 @@ impl LogPlot {
         F: FnOnce(&mut egui_plot::PlotUi),
     {
         plot_function(plot_ui);
-    }
-}
-
-fn offset_plot(plot: &mut PlotWithName, new_start_date: DateTime<Utc>) {
-    if let Some(first_point) = plot.raw_plot.first() {
-        let first_point_date = first_point[0];
-        let new_date_ns = new_start_date
-            .timestamp_nanos_opt()
-            .expect("Nanoseconds overflow") as f64;
-        let offset = new_date_ns - first_point_date;
-
-        log::debug!("Prev time: {first_point_date}, new: {new_date_ns}");
-        log::debug!("Offsetting by: {offset}");
-
-        for point in &mut plot.raw_plot {
-            point[0] += offset;
-        }
-    }
-}
-
-fn apply_offset_to_plots<'a, I>(plots: I, settings: &LogStartDateSettings)
-where
-    I: IntoIterator<Item = &'a mut PlotWithName>,
-{
-    for plot in plots {
-        if plot.log_id == settings.log_id {
-            offset_plot(plot, settings.start_date);
-        }
-    }
-}
-
-fn update_plot_dates(
-    percentage_plots: &mut [PlotWithName],
-    to_hundreds_plots: &mut [PlotWithName],
-    to_thousands_plots: &mut [PlotWithName],
-    settings: &mut LogStartDateSettings,
-) {
-    if settings.date_changed {
-        apply_offset_to_plots(percentage_plots.iter_mut(), settings);
-        apply_offset_to_plots(to_hundreds_plots.iter_mut(), settings);
-        apply_offset_to_plots(to_thousands_plots.iter_mut(), settings);
-
-        settings.date_changed = false;
     }
 }
